Never Back Down
Standing still as they charge,
We will hold our ground, not back down.
We will raise our swords, from the ground.
And we will fight the war to end,
All the suffering.
Never fearing heroes, foes
or the daring and their cause.
To the gods we pray, never waive.
Time has come for us to take charge
and fight the power.

Standing still as they charge,

We will hold our ground, not back down.
We will raise our swords, from the ground.
And we will fight the war to end,
All the suffering.
Never fearing heroes, foes
or the daring and their cause.
To the gods we pray, never waive.
Time has come for us, let's obey
and fight the power.