Everything Fades to Gray
Tonight my life will lack it's meaning
One final step before we're seeing
For once I am happy my little friend
Cannot comprehend what we are seeing

No words are needed, only a gaze, embrace
Contact the living, remember the dead

There is always something to learn,
Repetition proves it's worth
Without those moments, we would not remember

It's not fair, it's not fair, there was a time now
Kneeling down, take in the moment when
Everything becomes finally clear.
It's not fair to lose it,
And how, there was a time now
But Death is cheating us somehow

Come here now

Can't walk away, nothing to say
No need to feel so afraid,
Colours last a lifetime and fade to grey

It's not fair, it's not fair, there was a time now
Kneeling down, take in the moment when
Everything becomes finally clear.
It's not fair to lose it,
And how, there as a time now
But Death is cheating us somehow
And he's already here.

"When it all ends"
When everything fades to grey,
We dive into the darkness
Some things are needless to say

Somethings are needless to say,
We dive into the darkness
When everything fades to grey