Catechu dye
Start of some minor content to make things a bit easier

Catechu is an alternative to cocoa beans for brown dye
Oak apple galls are planned for black dye
Something might be planned for green dye. Unsure of what just yet.
Woad will be for blue dye in the second release to curse